+++ Bookong.Domain/Interfaces/IBookQrCodeImageGenerationService.cs
@@ -1,0 +1,7 @@
+namespace Bookong.Domain.Interfaces
+{
+    public interface IBookQrCodeImageGenerationService
+    {
+        byte[] GenerateQrCodeImage(Guid bookPublicId);
+    }
+}
