+++ Bookong.Infrastructure/Repositories/BookRepository.cs
@@ -4,9 +4,6 @@ using Bookong.Domain.Interfaces;
 using Bookong.Domain.Queries;
 using Bookong.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
-using QRCoder;
-using System.Drawing;
-using System.Drawing.Imaging;
 
 namespace Bookong.Infrastructure.Repositories
 {
@@ -185,18 +182,6 @@ namespace Bookong.Infrastructure.Repositories
             return await _context.Books
                 .Select(b => b.Id)
                 .ToArrayAsync();
-        }
-
-        public byte[] GenerateQrCodeImage(Guid bookPublicId)
-        {
-            using var qrGenerator = new QRCodeGenerator();
-            using var qrCodeData = qrGenerator.CreateQrCode(bookPublicId.ToString(), QRCodeGenerator.ECCLevel.Q);
-            using var qrCode = new QRCode(qrCodeData);
-            using var qrCodeImage = qrCode.GetGraphic(10);
-
-            using var ms = new MemoryStream();
-            qrCodeImage.Save(ms, ImageFormat.Png);
-            return ms.ToArray();
         }
     }
 }
